Fix GPLyness
Seeing as how I literally cut and pasted tons of mirrorselect code,
it is sensible/fair to go ahead and restore all the old authors
to the copyright lists.

Also, to ensure utter confusion/chaos, change the versioning system
up to a three-number system ("0.1.0") -- really should have put that
in a separate commit but lazy.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>