scripts/kernelng: position-independent arguments
Certain arguments should be respected regardless of
other arguments on the command-line.  Bundle them together
into a chunk of code and copy-paste it everywhere we
have a subcommand -- a nicer approach might be to augment
the knggroup/kngcommand decorators to do this for us but
so far I am too lazy.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>