eliminate boilerplate pattern
By moving the functional equivalent of the boilerplate into the
kng{command,group} decorators, we can get rid of the need for the
boilerplate in the first place.  The tradeoff is that the code is a lot
harder to read.  We could have made it easier to read by using the
decorator '@' syntax but this would have required an ugly null-decorator
to achieve.  So, I picked the better-performing kind of ugly.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>