kngclick: fix command-list highlighting
My first attempt to add the new yellow-bold styling for definition-list
terms worked for Options, but broke the white-bold styling I wanted for
the Commands definition lists.

To fix this, dl_style_word needs to know what section we are in; however,
this is not preserved anywhere accessible to dl_style_word.  To fix this,
we wrap the HelpFormatter.section() contextmanager to remember the
current section title before calling the inherited implementation, and
do the yellow-bold styling only if we are in the "Options" section.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>