kngclick: uglier, more effective section styling
This rewires the section styling by overriding write_heading and
heuristically guessing if the heading corresponds to the section.

If so, we do the styling (including, this time, a bold colon) there.
Also, add a kludge to de-pluralize "Commands" in the section header, for
clarity and consistency with the usage.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>