config: new file
Here we start building up a "config" framework to represent the state of
the kernel-ng configuration.  So far it is only able to write; as of yet
reading is not implemented.  Also implements the concept of "example"
and "default" intermediate configuration datastructures.  Kinda a
remnant of an aborted early attempt tbh... well, it's a start :/

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>