output: lazy trace formatting
As implemented, the trace decorators add a ton of useless overhead when
not in --debug mode, because they calculate all the formatting, and then
discard it when the trace level is too low.  Change the _at.say API to
expect a zero-argument lambda expression instead of a string, and only
invoke it when needed, to (hopefully) mostly fix this.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>