KNGTextWrapper: python2.7?
Mysteriously, I'm 75% sure this code was tested and working
under python 2.7.  I really can't explain that, so, perhaps I'm
simply mistaken or perhaps it's a recent 2.7 backport or something that
nobody bothered to document anywhere I could find it.

Regardless, attempting to read prose online, the clear consensus is that
the "(..., *, ...)" argument syntax for nameless positional arguments
is not supported in python 2.7 and never will be.

At the very least, it definitely was confusing ipython3.4.

So drop that & do the work the old-fashioned way (note: I'm still aiming
for 2to3-less python 2.7/3.x compatibility in this code-base).

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>