KNGConfig: cache globals property
There is no reason to clone a new globals instance every time we
touch this property -- they are all perfectly interchangeable as they
are virtual views onto singleton containees of the KNGConfig.

Therefore, construct one, the first time someone asks for it, and
return the same one from then on out.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>